data_catalog/requirements.txt
Django==1.6
django-tastypie==0.10.0
django-tastypie-elasticsearch==0.5.1
elasticsearch==0.4.3
ipython==1.1.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.4.1
urllib3==1.7.1
wsgiref==0.1.2
